LPS-85088 Expand ParameterMapUtilTest into individual test classes
